+++ manager2/src/app/app.module.ts
@@ -46,7 +46,7 @@ import { AdminpluginComponent } from './admin/adminplugin/adminplugin.component'
 import { FlashComponent } from './utils/flash/flash.component';
 import { TagComponent } from './utils/tag/tag.component';
 import { UserLogsComponent } from './user/userlogs.component';
-import {TableModule} from 'primeng/table';
+import { TableModule } from 'primeng/table';
 import { environment } from '../environments/environment';
 import * as Sentry from "@sentry/browser";