+++ routes/projects.js
@@ -581,6 +581,62 @@ router.delete('/pending/project/:uuid', async function (req, res) {
 
 });
 
+router.post('/pending/project/:uuid/reject', async function (req, res) {
+    if (!req.locals.logInfo.is_logged) {
+        res.status(401).send('Not authorized');
+        return;
+    }
+
+    const rejectionReason = req.body.rejectionReason?.trim();
+    if (!rejectionReason) {
+        res.status(400).send({ message: 'Rejection reason is required.' });
+        return;
+    }
+
+    const ownerEmail = req.body.rejectionReason?.trim();
+    if (!ownerEmail) {
+        res.status(400).send({ message: 'owner Email is required.' });
+        return;
+    }
+
+    let user = null;
+    let isadmin = false;
+
+    try {
+        user = await dbsrv.mongo_users().findOne({ _id: req.locals.logInfo.id });
+        isadmin = await rolsrv.is_admin(user);
+    } catch (e) {
+        logger.error(e);
+        res.status(404).send({ message: 'User session not found' });
+        return;
+    }
+
+    if (!user) {
+        res.status(404).send('User not found');
+        return;
+    }
+    if (!isadmin) {
+        res.status(401).send('Not authorized');
+        return;
+    }
+
+    try {
+        await maisrv.send_notif_mail({
+            name: 'rejection_project_creation',
+            destinations: [ownerEmail, CONFIG.general.accounts],
+            subject: `[Creation project rejected] `,
+            message: `Your project "${ownerEmail}" has been rejected for the following reason: ${rejectionReason}`,
+        });
+        logger.info(`Rejection email sent to: ${ownerEmail}`);
+    } catch (emailError) {
+        logger.error('Failed to send rejection email:', emailError);
+        res.status(500).send({ message: 'Failed to send rejection email.' });
+        return;
+    }
+
+    res.send({ message: 'Rejection email sent successfully.' });
+});
+
 router.get('/project/:id/users', async function(req, res){
     if(! req.locals.logInfo.is_logged) {
         res.status(401).send({message: 'Not authorized'});