+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.4.32 (Unreleased)
 
+* New project: make description field required
 * Clearer error message when adding a website
 * Add User, Project and Group static classes to front end (refactor)
 * Add checks for website owner update